CC-5321 Float stock revert

+++ Bundles/Oms/src/Spryker/Zed/Oms/Business/OmsFacade.php
@@ -314,7 +314,7 @@ class OmsFacade extends AbstractFacade implements OmsFacadeInterface
      * @param string $sku
      * @param \Generated\Shared\Transfer\StoreTransfer|null $storeTransfer
      *
-     * @return float
+     * @return int
      */
     public function sumReservedProductQuantitiesForSku($sku, ?StoreTransfer $storeTransfer = null)
     {
@@ -331,7 +331,7 @@ class OmsFacade extends AbstractFacade implements OmsFacadeInterface
      * @param string $sku
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
      *
-     * @return float
+     * @return int
      */
     public function getOmsReservedProductQuantityForSku($sku, StoreTransfer $storeTransfer)
     {
@@ -634,11 +634,11 @@ class OmsFacade extends AbstractFacade implements OmsFacadeInterface
      *
      * @param string $sku
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
-     * @param float $reservationQuantity
+     * @param int $reservationQuantity
      *
      * @return void
      */
-    public function saveReservation(string $sku, StoreTransfer $storeTransfer, float $reservationQuantity): void
+    public function saveReservation(string $sku, StoreTransfer $storeTransfer, int $reservationQuantity): void
     {
         $this->getFactory()
             ->createUtilReservation()
